backend/requirements.txt
Django~=3.0.6
djangorestframework
markdown
django-filter